[8] convert TCP/UDP ports into protocol names and vice versa
This commit modifies the type PT_PORT to accept both a numeric
value and a string value specified in a services file.

It adds a services file from wireshark and code to parse it from
NetBSD.  The code that does this is being maintained with tests
outside of this repository at

git@github.com:kristopolous/services-parser.git

It introduces a config.h inside of libsinsp in order to find
where the services file will live during runtime.  The copying is
done inside libsinsp's Cmakefile.

The servparse.h and .cpp introduce 4 functions to convert to and
from numeric ports and string values.