driver: Make add_sentinel a "real" inline function
Putting an inline function in a .c file does not do anything except
confuse the compiler.  Newer versions of gcc will complain loudly about
this, and refuse to build, so put the function in a .h file so that it
can actually be used as an inline function, making everyone happy.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>