driver: properly predefine variables
Two variables were defined in ppm.h, but the .c files that declared them
did not include ppm.h, making sparse complain about these.  Properly
include the .h file, and move 2 structures that are only used in main.c,
into main.c to avoid having to have a bunch of .h files in ppm.h that
are not needed.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>