[8] Split of /etc/services in a _WIN32 block
This commit pretends to support windows version of /etc/services
by using an IFDEF _WIN32 block and then pointing to

	\\windows\\system32\\etc\\services
	^                   ^
	+-------+-----------+
I'm aware that  `-> this block right here is totally configurable
based on how windows is installed, and that there's probably some
win32 routine that i've long forgotten about to get the %SYSTEM%
directory and inject it in there.

I defer to the windows gods to look over and put that in - or
to punt it back to me and tell me to figure it out myself (;0)).

Alternatively, we can just drop support for the OS level services
in windows if that would make things easier overall.  It was totally
out of any spec and just done as me being mr. fancypants.