syscall table no longer checks for x86_64. Looks at the individual syscalls
There were some __x86_64__ checks whose purpose wasn't entirely clear. Mostly I
THINK they were meant to separate the syscalls that existed on each
architecture, but this wasn't done perfectly: there were syscalls that existed,
but were masked out by the #ifdef. This patch removes the check for x86_64, and
checks each syscall for existence individually.

It may be a good idea to do this across the board for ALL the syscalls. The C
preprocessor can't quite do that I don't think, but we can generate the header
we want.