[71] Detect in-tree build and fail
This commit will check to see if the file CMakeLists.txt exists in
the current directory and if so, will assume that the user is trying
to build in tree.

It will immediately fail and print out instructions on how to properly
build sysdig along with a shortened link to the wiki page for more
information.