Outputs JSON capture objects in a large array
This code introduces an m_first bool to the eventformatter that
represents if its the first tostring call. This is used in order to
emit a '[' or a ',' for a start of an array, or the array delimeter,
respectively.

Also, sinsp_evt_formatter::on_capture_end is introduced, which, for
JSON will emit the trailing ']' representing the end of an array.

This means that an extra clause was put into the g_terminate conditional
inside of do_inspect which will do the on_capture_end of the inspector,
thus emitting the trailing ']'.

Additionally, the JSON::Value and JSON::FastWriter have been moved
to the class scope.