Fixed line-wrapping when printing a chiesel arg list
Let's say I have a chisel test.lua with an argument list defined like this:

 args =
 {
         {
                 name = "arg1",
                 description = "blah blah blah blah blah",
                 argtype = "int"
         },
         {
                 name = "arg2",
                 description = "something",
                 argtype = "int"
         },
 }

Then sysdig -i test prints the argument list like this:

 Args:
 [int] arg1 - blah blah blah blah blah
 [int] arg2 - so
                     mething

Note the incorrect line wrapping. This was due to a bug where the wrapping was
applied to the whole string, not just a line at a time. Resolved in this patch