install a rebuilt version of the manpage if pandoc is available
sysdig.8 is generated from sysdig.md. When a packager wants to modify
the manpage, he has to modify the manpage and not the source, which is
wrong. Let's rebuild the manpage if pandoc is available.

In a perfect world, we would get rid of the prebuilt version at all,
but pandoc may not be available on some tagets.