[133] Implements Additional options/flags (mimic tcpdump -C file_size)
This commit introduces an engine called the "cycle writer" that is
intended to mimic tcpdump's style of captures spanning multiple files.

It supports the same arguments as tcp dump and has the same
functionality.  It was implemented independently, without viewing
any of the tcpdump code and so should be free to fall under any license.