Generate code for events and flags from declarative description.
Script generate.lua is used to generate C code used to define
properties of events and flags, which is scattered throughout sysdig
codebase, from input files flags.lua and events.lua.

Goals of this patch are:
[*] Make addition of new events (syscalls) simple, intuitive, and less
    error-prone
[*] Catch syntax and semantic errors earlier (sensible assertion message
    from Lua generator vs C compiler errors because of misplaced '}')
[*] Avoid code duplication between driver and userspace.
[*] Simplify maintainance of event defintions.

Most of events and half of flags are not ported to new infrastructure yet.

See discussions in [1,2] for the background of this patch.

[1] https://github.com/draios/sysdig/pull/180
[2] https://groups.google.com/forum/#!topic/sysdig/u4q3FTRasE8