cpu hotplug: handle new cpus being added at any point in time.
If a cpu is removed, we don't free up the memory for it, hoping it will
come back sometime soon (which is usually the case.)  Note, now all cpus
possible are shown as device nodes in the system, if you try to open one
for a cpu that is not present, open will properly fail.  This might
affect userspace tools that expect to be able to watch for all cpus in
the system.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>