chisels: Add parse_numeric_input() in common.lua
This commit adds a function in common.lua to parse numeric user input
and exit with a nice error message if something goes wrong.

This new function uses tonumber() in order to convert user input to a
number. If tonumber() returns 'nil', aka input wasn't a number, the
function prints an error message to inform the user and calls
os.exit() to stop execution.

All chisels that need to accept numeric user input should use this
function from now on.