Add new and extend existing syscalls. Add `evt.abspath` filter field.
This commit adds support for the following syscalls:

- symlink
- symlinkat

and extends already existing syscalls by transfering their parameters
to the exit handler. See the following link for motivation:

https://github.com/draios/sysdig/pull/246#issuecomment-55964233

For the time being only the exit handler is extended since changing to
entry handler as well would break older chisels.

In addition to those syscalls a new set of filter fields was added:

- evt.abspath
- evt.abspath.src
- evt.abspath.dst

Those field are only used for the syscalls `renameat`, `openat`,
`linkat`, `unlinkat` and `symlinkat` and aggregate the values of
corresponding `dirfd` and `name` parameters to form an absolute path.