Add `symlink(at)?` syscalls. Add `evt.abspath` filter field.
This commit adds support for the following syscalls:

- symlink
- symlinkat

In addition to those syscalls a new set of filter fields was added:

- evt.abspath
- evt.abspath.src
- evt.abspath.dst

Those field are only used for the syscall `symlinkat` but might be
later used for other `...at` syscalls as well (e.g. `renameat`,
`openat`, `linkat` and `unlinkat`). The purpose of the filter fields
is to aggregate the values of corresponding `dirfd` and `name`
parameters to form an absolute path.