Transfer params to exit event by providing alternative syscall handlers.
This commit adds additional event pairs for the following syscalls:

- openat
- link
- linkat
- unlink
- unlinkat
- mkdir
- rmdir

See the following link for motivation:

https://github.com/draios/sysdig/pull/246#issuecomment-55964233

The general pattern for those new event was to take all parameters out
of the entry event and append them to the exit event instead.

To keep backwards compatibility with older chisels this commit mixes
existing and new events in the `syscall_table.c`. For the time being
all syscalls mentioned above generate their "old" entry event and their
"new" exit event. That allows access to all parameters to be available
on both entry and exit events.