Fix check_id setting on events
Before this fix, the wrong check_id could be set on an event which
passed a filter.

For example in `A OR B` where A evaluates to true and B evaluates to
false, the check_id from B would incorrectly be set on the event.

This fix addresses that (in a cleaner way than the earlier fix on the
`demo-hack` branch).