Skip additional syscalls/events for falco.
We found that falco still had significant cpu usage on production
servers with a lot of activity, so do another round of looking for
syscalls/events that can be skipped for falco. That list includes:

 - futex
 - stat, lstat, stat64, lstat64, fstat64
 - statfs, fstatfs, statfs64, fstatfs64, fstatat64
 - lseek, llseek
 - mmap, mmap2, munmap
 - procinfo
 - sched_yield
 - ioprio_get, ioprio_set

Also the syscall socket was not reenabled as a part of 9c489b3, so
reenabling here.