Gracefully exit on large read buffers.
Gracefully exit if the size of an event is larger than the size of the
read buffer (currently 64k). Although events could theoretically be
larger given the number of parameters and size of each parameter, in
practice no event is really that big.

If this happens it returns and error and exits.