Start tracking session ids for threads.
Start tracking session ids in threads and add the ability to filter on
session ids in filter expressions.

Modify scap_threadinfo/sinsp_threadinfo to add a sid (session id)
member. When filling in details on a process in
scap_proc_fill_info_from_stats(), the session id is read from the
contents of /proc/<pid>/status, which was already being read--we just
needed to fill in the value from the sscanf().

The sid is written and read along with other thread information in trace
files. This also bumps the block header version from V5 to V6. In places
where the block header was checking for versions <= 5, they needed to be
updated to <= 6.

A new filter expression proc.sid matches/displays the thread's session
id.

On a clone(), the session id is transferred to the new thread from the
parent. The same restrictions apply as the parent pid--if information on
the parent process is not available for some reason, the session id will
not be set.

The session id is not currently updated on a setsid() call. That will be
handled in the next commit.