Handle cases where the session leader has exited.
When resovlving proc.sname, handle the case where the process with
pid=proc.sid has exited. I ran into this when using the pipe installer
example, where the top level bash installer spawns subshells which in
turn run the python, etc processes we want to track.

In this case, walk the process ancestor heirarchy until the session id
changes or you reach the top process. The top process having the same
session id is considered the process leader, and its comm is used for
proc.sname.