Add regression tests for proc.sid/proc.sname.
The actual new change is to add a second trace file sessions.scap which
shows the results of calling "setsid -w ls", with results scoped to the
shell where setsid is run. This handles all the cases of tracking a
session id/name for an existing process (the base shell), execing
"setsid -w ls", which changes the sid and sname, and waiting for the ls
to exit.

In order to add this new trace file, I needed to change how the
regression tests were run. The current regression tests assumed that for
a given input trace file that there would be a corresponding baseline
directory with the results of running that trace file with all the sets
of arguments in sysdig_trace_regression.sh. That didn't really handle
forward-looking cases like this, where I don't want to run old files
with new combinations of arguments and didn't want to use old arguments
on the new trace file.

Fix this by changing sysdig_batch_parser to skip running sysdig with a
given set of arguments when the corresponding baseline directory doesn't
exist. If the directory does exist but the individual baseline file
doesn't exist, skip that file.

This change has zip files baseline-add-session-tests.zip and
traces-add-session-tests.zip, which should work based on my
understanding of TRAVIS_BRANCH. After merging this branch to dev I'll
update the actual baseline.zip/traces.zip files, and give anyone working
on an active branch to rebase.

This won't handle the case of the master branch, though. I need to check
to see if sysdig_trace_regression is run outside of travis.