Add support for setsid events.
Add support for setsid events, which allows for detailed information on
the arguments/return value as well as parsing to update the session id
on a successful setsid call.

Add the new events PPME_SYSCALL_SETSID_{E,X} to event_table both for the
driver and libscap, and add the events to the syscall table for
setsid. For exit events, the new session id is filled using
PPM_AUTOFILL.

In libsinsp, on setsid exit events, when the return value is
successful (>= 0) update the thread's session id m_sid to the return
value of the setsid call.