Add startswith as a string comparison operator.
Currently, the only way to do a substring match is the 'contains'
operator. This is less than ideal when you specifically want to test the
beginning of a string for the substring, as contains will traverse the
entire string and may return a match in the middle of the string.

Add a startswith operator that specifically tests for a substring match
at the beginning of the string.

There were a few places in non-string comparison operators where any
operator not in the set of supported operators was assumed to be
'contains'. Replace this with checks for all operators including
'icontains', 'startswith' with appropriate error messages.