Make sanitize_string cfgable and off for filters.
Modify sinsp_filter_check::extract to take an optional argument
sanitize_strings, defaulting to true. It controls whether the
filtercheck should sanitize any string arguments when extracting the
value from the event.

The only place where it is set to false is in
sinsp_filter_check::compare(). This means that when comparing events
against a filter expression, any string values are not sanitized. They
will still be sanitized when the events are displayed.

This significantly speeds up users like falco that have complicated
filter expressions that run against lots of events.