Dont sanitize strings filters (#625)
* Switch back to erase/remove_if for sanitization.

It turns out that with -O3 (release flags), using erase/remove_if with
an operator struct is slightly faster than the inline version that was
here, as the compiler can inline the operators. It isn't faster with -O2
and is much slower without optimization.

Optimize for the release build by adding the operators back.

* Whitespace diffs.

Making whitespace changes in separate commit.

* Make sanitize_string cfgable and off for filters.

Modify sinsp_filter_check::extract to take an optional argument
sanitize_strings, defaulting to true. It controls whether the
filtercheck should sanitize any string arguments when extracting the
value from the event.

The only place where it is set to false is in
sinsp_filter_check::compare(). This means that when comparing events
against a filter expression, any string values are not sanitized. They
will still be sanitized when the events are displayed.

This significantly speeds up users like falco that have complicated
filter expressions that run against lots of events.