Changes for falco-engine (#640)
* Minor header file cleanups.

Related to changes in https://github.com/draios/falco/pull/103. Probably
best to let that PR merge first.

* Add abilty to disable filters.

Most of the changes are to pass along a name with a filter when adding a
filter via add(). This name can be used for a new method
sinsp_evttype_filter::enable(), which takes a pattern regex which is
matched against the names and a bool enabled flag.

To manage the new name and enabled state along with the filter, wrap the
existing sinsp_filter object in a wrapper filter_wrapper containing the
sinsp_filter, enabled flag, and name. Everywhere a sinsp_filter object
was used, a filter_wrapper is used instead.

The list of filters is now a map from name to filter wrapper.

Enable just loops over the filters, matching the names against the
pattern, and toggles the enabled flag for those filters that match.

In run(), when looping over filters check the enabled flag before
running the filter.