Ensure PT_CHARBUF extracts set length.
Ensure that all PT_CHARBUF filtercheck extracts set the length of the
string they return. Although the strings they return are usually
null-terminated, they may not always be null terminated, which causes
problems in the chisel api when returning the strings to lua.

This is half of the fix for https://github.com/draios/sysdig/issues/736.