Minimal support for NETLINK sockets.
Add minimal support for AF_NETLINK sockets, with the new type
SCAP_FD_NETLINK. This is mostly done so they are not mixed up with
SCAP_FD_UNKNOWN sockets, which are not expected to be in scap files at
all.

The type is used when first parsing a fd from its fd num/domain num in
sinsp_parser::add_socket or when parsing /proc/net/netlink in
scap_fd_read_netlink_sockets_from_proc_fs(). After that, the type is
copied when converting from scap_fdinfos to sinsp_fdinfos in
sinsp_threadinfo::add_fd_from_scap.

There is limited support for displaying information on netlink fds. The
typechar is 'n', the typestring is "netlink"/"<NETLINK>", and they don't
have any additional information or a category for them.