Improve handling of potentially unknown fds.
Ensure we better handle fds that can't be mapped to a specific
scap_fd_type.

When reading a descriptor from a scap file, ignore SCAP_FD_UNKNOWN
sockets and don't return an error. When writing a descriptor to a scap
file, skip SCAP_FD_UNKNOWN descriptors.

When parsing a socket fd, if we don't map a type log an error but continue.