Adding ability to open trace files from fds (#838)
* Adding ability to open trace files from fds

Add the ability to open scap files or inspectors from a fd. Makes it
possible to open other objects that return a fd and treat those objects
as if they were a file.

 - Add new functions scap_open_offline_fd, scap_dump_open_fd that take
   all the same arguments as their filename versions but take a fd
   instead of a filename. scap_open_offline_int now takes a gzfile which
   has been opened by one of the top-level functions.
 - scap_open's args struct now contains a fd. If non-zero, it takes
   precedence over filename and the fd is used to open a gzfile via
   gzdopen().
 - New function scap_event_reset_count sets the read event count to
   0. Used when rewinding a scap file to start over.
 - Add some more useful error messages to
   scap_read_{proclist,userlist,fdlist} when reading more than expected.
 - also add fdopen variants to sinsp_dumper/sinsp. They call the
   appropriate scap fd variants. most of sinsp::open moves to sinsp::open_int().
 - Add is_open() and written_events() methods to sinsp_dumper(). Useful
   when writing files via fd to track progress.
 - In sinsp::init(), after rewinding the file also reset the event
   count.
 - Add a sinsp::fseek() which just uses scap_fseek().

* Add ftell/ability to read bytes written.

Add scap_dump_ftell, which maps to gztell, and
sinsp_dumper::next_write_position, which uses scap_dump_ftell.

* Add ability to skip proc scan for fd opens.

Add an argument to scap_dump_open_fd that controls whether or not to do
a scan of /proc when initializing the dump file.

In sinsp_dumper::fdopen, set this to false. Also write the thread table
to the file, following recent changes to sinsp_dumper::open.