Set close-on-exec flag for driver fds.
While debugging some agent unit tests, I noticed that after a fork +
exec the number of driver references was higher than expected. Some of
these problems can be fixed within the tests, but it's also useful to
set the close-on-exec flag for the driver fd so it's closed during an
exec.

Also fix some error messages that were referring to a hardcoded
sysdig_probe instead of PROBE_DEVICE_NAME.