libsinsp: rename filter `proc.nchilds` to `proc.nchildthreads`
Change the implementation of `proc.nchilds` to respect what's written in
the documentation. Before this patch, proc.nchilds was extracted from
the thread, instead that from the process. With this patch the only
difference between `proc.nthreads` and `proc.nchilds`(aka
`proc.nchildthreads`) is that the former includes the main thread in the
result, the latter doesn't.