Change initial read to byte limit + increase (#892)
In get_all_data(), which is used when initially reading k8s environment
data, there used to be a limit on the number of reads performed. This
was a bit sensitive to the packet size and delivery pattern, and in some
cases could give up after as few as 50-100k.

Change this to a limit on the number of bytes read, and change the limit
to 30mb.