Return container meta events.
Previously, container/k8s/mesos meta events, which included information
about a newly created container/k8s/mesos entity, were dumped directly
to capture files but not returned in sinsp::next().

This causes problems if you're using a sinsp_dumper outside of the
inspector and want to dump events to a file yourself.

To fix this, add a new inspector option set_internal_events_mode which
controls whether or not internal events (i.e. container, k8s, mesos) are
returned in live mode.

In the sysdig program, the -D option also enables internal events mode.

Also ensure that container meta events are returned/written before the
event that creates the container. This involves saving the current scap
event to m_meta_skipped_evt when a meta event is created as a
side-effect of sinsp_parser::process_event(), and returning that event
the next time sinsp::next() is called. With this change, there isn't any
need for special handling of container events when dumping, so remove
that code.