Overwrite ptid during execve (#914)
In some scenarios we are missing many clone() events, but not the relative execve() (in the process of troubleshooting the cause of this).

In those cases, whenever the first event for the newly created process happens, sysdig will create a new fake "<NA>" process that doesn't have most fields set. Whenever the execve happens, some of these fields will be set (args, comm, container_id, ...), however the ptid remains to the default `-1`, meaning that parent navigation will not work.

From the point of view of sinsp, this process will then appear as being inside a container but detached from the main process hierarchy, and this creates some issues on the Monitor/Secure side of things.