Don't use unrelated m_meta_evt for k8s/mesos evts
When creating k8s/mesos events, it used to use m_meta_evt, which was a
not-really-related event that happened to be the last container meta
event that was created. There isn't any guarantee that that event has
anything to do with the k8s/mesos event being created, or is even valid.

This all happened to work because the only thing being used from
m_meta_evt was the thread id, which is always 0 (see
https://github.com/draios/sysdig/blob/dev/userspace/libsinsp/container.cpp#L538
and
https://github.com/draios/sysdig/blob/dev/userspace/libsinsp/container.cpp#L588).

So directly just set the thread id of the k8s/mesos event to 0 and get
rid of the connection.