Get rid of unused m_skipped_evt.
m_skipped_evt is not actually used, and might be confused with
m_meta_skipped_evt, which is a system event like execve, clone, etc. is
delayed when a container is created and emitted/written before the
system event.

Get rid of m_skipped_evt and the unused method add_meta_event_and_repeat
which used it.