Return container meta events. (#912)
* Return container meta events.

Previously, container/k8s/mesos meta events, which included information
about a newly created container/k8s/mesos entity, were dumped directly
to capture files but not returned in sinsp::next().

This causes problems if you're using a sinsp_dumper outside of the
inspector and want to dump events to a file yourself.

To fix this, add a new inspector option set_internal_events_mode which
controls whether or not internal events (i.e. container, k8s, mesos) are
returned in live mode.

In the sysdig program, the -D option also enables internal events mode.

Also ensure that container meta events are returned/written before the
event that creates the container. This involves saving the current scap
event to m_meta_skipped_evt when a meta event is created as a
side-effect of sinsp_parser::process_event(), and returning that event
the next time sinsp::next() is called. With this change, there isn't any
need for special handling of container events when dumping, so remove
that code.

* Don't use unrelated m_meta_evt for k8s/mesos evts

When creating k8s/mesos events, it used to use m_meta_evt, which was a
not-really-related event that happened to be the last container meta
event that was created. There isn't any guarantee that that event has
anything to do with the k8s/mesos event being created, or is even valid.

This all happened to work because the only thing being used from
m_meta_evt was the thread id, which is always 0 (see
https://github.com/draios/sysdig/blob/dev/userspace/libsinsp/container.cpp#L538
and
https://github.com/draios/sysdig/blob/dev/userspace/libsinsp/container.cpp#L588).

So directly just set the thread id of the k8s/mesos event to 0 and get
rid of the connection.

* Get rid of unused m_skipped_evt.

m_skipped_evt is not actually used, and might be confused with
m_meta_skipped_evt, which is a system event like execve, clone, etc. is
delayed when a container is created and emitted/written before the
system event.

Get rid of m_skipped_evt and the unused method add_meta_event_and_repeat
which used it.

* Make notification events internal.

This allows them to be treated identically to other internal events in
capture files.