Allow proc info to be MAX_*_SIZE in length
One of the changes in https://github.com/draios/sysdig/pull/845 was to
refuse to read process information that was exactly SCAP_MAX_*_SIZE
characters in length, to make sure that the resulting string fit in a
buffer of size SCAP_MAX_*_SIZE, including the trailing null.

This caused a problem when the trace file contained a process
information string that is exactly SCAP_MAX_*_SIZE characters long,
which is actually allowed in the trace file.

The fix is to increase the buffer size by 1 and read up to
SCAP_MAX_*_SIZE characters, followed by setting array[size] to '\0'.