Fix more valgrind errs (#938)
* Initialize exepath to empty string.

When converting a sinsp_threadinfo to a scap_threadinfo, make sure
exepath is initialized.

This change was originally in
https://github.com/draios/sysdig/pull/845/files, but got dropped during
a conflict merge. This adds the change back.

* Add scap_threadinfo ctor and use it vs ad-hoc init

Valgrind was reporting memory leaks when reading/writing trace files,
and the leak was the table scap_threadinfo::scap_fdinfo as used by
sinsp_thread_manager::dump_threads_to_file().

This function was re-using the same scap_threadinfo struct for all
processes. The individual fds were being cleaned up, as they were in the
vector fdinfos_to_free, but not the hash table that contained them.

Fix this by reorganizing dump_threads_to_file a bit to not reuse a
scap_threadinfo struct across all the processes, but to use individually
created ones for each process, using scap_proc_free() to destroy the
scap_threadinfo struct.

There was no corresponding constructor, so add one that takes care of
the malloc/memset as well as setting handle->m_lasterr when the malloc
fails. Use that in dump_threads_to_file as well as in
scap_proc_add_from_proc().

* Add ability to set dumper's inspector.

This is used in other products that keep a sinsp_dumper around with
different inspectors.

* Allow proc info to be MAX_*_SIZE in length

One of the changes in https://github.com/draios/sysdig/pull/845 was to
refuse to read process information that was exactly SCAP_MAX_*_SIZE
characters in length, to make sure that the resulting string fit in a
buffer of size SCAP_MAX_*_SIZE, including the trailing null.

This caused a problem when the trace file contained a process
information string that is exactly SCAP_MAX_*_SIZE characters long,
which is actually allowed in the trace file.

The fix is to increase the buffer size by 1 and read up to
SCAP_MAX_*_SIZE characters, followed by setting array[size] to '\0'.