libsinsp: use global_pid when clone exit in the father
Before this patch, when a new thread was created, if the thread was
inside a container, only the child clone-exit event was able to
correctly create the thread, because it was the only one to have the
right tid. Instead the father had the pid of the current namespace.

Since now the tid of the init namespace is exported together with kernel
events, we can leverage that feature to immediately create the thread
from the parent.

In particular, this patch avoids inconsistencies when CLONE_PARENT flag
is used in `clone`.