Use explicit Docker API version for container metadata
As of this writing, Docker recommends at least v1.24, so we use that.
However, for e.g. RHEL 6/CentOS 6, which are stuck on Docker 1.7
with API version 1.19, we do a fallback request without the version
number. This way, we use v1.24 if possible, and the newest supported
version if not.

Closes: #900