Add ability to store values with search paths (#1007)
Previously, path_prefix_search was a set membership test. You could test
a path P against a set of search paths S, and see if any Si was a prefix
of P.

Extend that to also store values along with each search path Si, making
it a map path_prefix_map instead of a set. This is done by templating
the class with a Value v and storing the value at each leaf node in the
tree. When searching, when you find a leaf node a pointer to its
associated value is returned.

path_prefix_search remains as an instantiation of the templated map with
dummy values.