Also emit all mesos-related err logs as user evts
In an attempt to improve visibility without having to ask customers for
logs, also emit all mesos-related errors as user events by passing
through the json error log module.

As much as possible, the document is still mapped to the first argument
and the url is mapped to the last argument. When not possible,
placeholders are used.