Mesos custom events (#1009)
* Copy token_bucket from falco to sysdig

Falco has a good token bucket implementation and we want to use it in
sysdig. So copy the one from falco to libsinsp.

In a subsequent PR we'll remove the falco copy.

* Emit custom events for json parse errors.

Modify json_error_log::log() to also pass a uri (presumably the uri that
resulted in the json) and timestamp. This is used to build a custom user
event that contains the following:
  - event_name: "json_parse_error"
  - desc: the error string passed to log()
  - scope: {host.mac: machine id}
  - tags:
     - source: "json_parser"
     - uri: the provided uri
     - json_prefix: the first 100 chars of the json document

User events are throttled by a separate token bucket which limits the
rate of user events to once per 5 minutes and a burst of 10
events. (This is more aggressive than the overall user event
throttling).

In order to provide a uri, mesos_http::try_parse now passes around a
uri, which is usually the uri that was fetched but can also be strings
like "fixed-mesos-state" or "parse-mesos-evt".

* Also emit all mesos-related err logs as user evts

In an attempt to improve visibility without having to ask customers for
logs, also emit all mesos-related errors as user events by passing
through the json error log module.

As much as possible, the document is still mapped to the first argument
and the url is mapped to the last argument. When not possible,
placeholders are used.

* Add marathon errors, some addl mesos warnings

Also look in all marathon code for SEV_ERROR logs and emit them as user
events.

And also emit some selected warning log messages that we know are
especially interesting.